Add custom hook for error handling
